+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Remove old copyright & license fields.
 - Clean up KiCad project file.
 
+### Changed
+
+- Remove hard-coded project metadata from tasks.py.
+
 ## [2.0.1] - 2025-05-12
 
 ### Added